requirements.txt
# Devkit Python 依赖（所有服务的统一来源）
# 安装: .venv/bin/pip install -r requirements.txt

# ── 核心服务 ──────────────────────────────────
fastapi
uvicorn[standard]
aiohttp
python-multipart

# ── Gateway 客户端 ────────────────────────────
websockets
cryptography

# ── 图数据库（任务图系统）────────────────────
neo4j
pydantic

# ── Telegram Bot ──────────────────────────────
python-telegram-bot

# ── 科研工具 ──────────────────────────────────
pandas
numpy
scipy
matplotlib
paperscout
arxivy
dblpcli
s2cli
papis

# ── 系统监控 ──────────────────────────────────────
psutil

# ── 文档语义搜索 (RAG) ────────────────────────────
llama-index-core
llama-index-embeddings-openai

# ── Web / GUI 自动化 ─────────────────────────
# playwright  (uncomment when needed)

# ── Web Push 通知 ─────────────────────────────
pywebpush

# ── Markdown → TTS 纯文本 ─────────────────────
mistune

# ── 定时与周期调度 ─────────────────────────────
croniter

# ── 日历 / 通讯录 / 智能家居 ─────────────────
khal
vdirsyncer
khard
homeassistant-cli

# ── 农历转换（CalendarChecker 用） ────────────
lunardate
